+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Version 2.0
+- Support for N output colors (3 MIDI notes per color, default: 2)
+- Greatly improved screen capturing performance using vlcj library
+
 ### Version 1.1.0
 - Improved/simplified color calculation/normalization
 - Added file-lock to prevent multiple simultaneous instances